Move app logic to separate structure
