+++ .gitignore
@@ -2,7 +2,7 @@ __debug_bin
 server.key
 server.crt
 dist/
-uncors
+./uncors
 .idea
 node_modules
 .uncors.yaml
